requirements.txt
# P2P Swap Bot - Core Dependencies
# Minimal requirements for production deployment

python-telegram-bot==21.3
bitcoinlib==0.7.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyyaml==6.0.2
